+++ extensions/src/city/city_zone.cpp
@@ -22,10 +22,10 @@
  SOFTWARE.
  */
 
-#include <city/city_zone.h>
+#include <City/city_zone.h>
 #include <engine/engine.h>
-#include <city/city_map.h>
-#include <city/city_building.h>
+#include <City/city_map.h>
+#include <City/city_building.h>
 
 namespace neko
 {
@@ -61,7 +61,9 @@ void CityZoneManager::UpdateZoneTilemap(const CityBuilderMap& cityMap, const Cit
             continue;
         //culling with mainView
         const auto zoneSize = sf::Vector2f(tileSize);
-        const auto worldPos = sf::Vector2f(zone.position.x * tileSize.x, zone.position.y * tileSize.y);
+        const auto worldPos = sf::Vector2f(
+			float(zone.position.x * tileSize.x), 
+			float(zone.position.y * tileSize.y));
         const sf::FloatRect tileRect = sf::FloatRect((worldPos - zoneSize/2.0f), zoneSize);
 
         if (!windowView_.intersects(tileRect))
